requirements.txt
asgiref==3.2.7
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
dj-webhooks==0.2.1
Django==3.0.5
django-crispy-forms==1.9.0
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-rq==2.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
rq==1.3.0
six==1.14.0
sqlparse==0.3.1
standardjson==0.3.1
urllib3==1.25.9
webhooks==0.4.2
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
